Moves Mock to test zones
The Mock structure, Mock module and the Service class logic handling
Mocks are now into lib/fog/test_helpers where they can be reused by
other fog-<providers>.

SCP and SSH mock classes have also been moved to spec/ as they aren't
part of the Mock structure itself but just consuming it.

https://github.com/fog/fog-core/issues/198
