+++ lib/fog/core.rb
@@ -40,7 +40,6 @@ require File.expand_path('../core/errors', __FILE__)
 require File.expand_path('../core/hmac', __FILE__)
 require File.expand_path('../core/logger', __FILE__)
 require File.expand_path('../core/model', __FILE__)
-require File.expand_path('../core/mock', __FILE__)
 require File.expand_path('../core/provider', __FILE__)
 require File.expand_path('../core/service', __FILE__)
 require File.expand_path('../core/ssh', __FILE__)
