+++ lib/fog/core/service.rb
@@ -20,10 +20,6 @@ module Fog
         service.collections
       end
 
-      def mocked_requests
-        service.mocked_requests
-      end
-
       def requests
         service.requests
       end
@@ -51,42 +47,36 @@ module Fog
 
       # {Fog::Service} is (unfortunately) both a builder class and the subclass for any fog service.
       #
+      # A Service class is called Real because there is a mock equivalent class +Mock+ that can be used
+      # for tests. They are located under same tree structure starting from test/lib/fog.
+      #
       # Creating a {new} instance using the builder will return either an instance of
-      # +Fog::<Service>::<Provider>::Real+ or +Fog::<Service>::<Provider>::Mock+ based on the value
-      # of {Fog.mock?} when the builder is used.
+      # +Fog::<Service>::<Provider>::Real+
       #
       # Each provider can require or recognize different settings (often prefixed with the providers
       # name). These settings map to keys in the +~/.fog+ file.
       #
       # Settings can be passed as either a Hash or an object that responds to +config_service?+ with
-      # +true+. This object will be passed through unchanged to the +Real+ or +Mock+ service that is
-      # created. It is up to providers to adapt services to use these config objects.
+      # +true+. This object will be passed through unchanged to the +Real+ service that is created.
+      # It is up to providers to adapt services to use these config objects.
       #
-      # @abstract Subclass and implement real or mock code
+      # @abstract Subclass and implement real
       #
-      # @param [Hash,#config_service?] config
+      # @param [Haservicesh,#config_service?] config
       #   Settings or an object used to build a service instance
       # @option config [Hash] :headers
       #   Passed to the underlying {Fog::Core::Connection} unchanged
       #
-      # @return [Fog::Service::Provider::Real] if created while mocking is disabled
-      # @return [Fog::Service::Provider::Mock] if created while mocking is enabled
+      # @return [Fog::Service::Provider::Real] if created
       # @raise [ArgumentError] if a setting required by the provider was not passed in
       #
       # @example Minimal options (dependent on ~/.fog)
       #   @service = Fog::Compute::Example.new # => <#Fog::Compute::Example::Real>
       #
-      # @example Mocked service
-      #   Fog.mock!
-      #   @service = Fog::Compute::Example.new # => <#Fog::Compute::Example::Mock>
-      #
       # @example Configured using many options (options merged into ~/.fog)
       #   @options = {
       #     :example_username => "fog",
-      #     :example_password => "fog"
-      #   }
-      #   @service = Fog::Compute::Example.new(@options)
-      #
+      #     :examplservice
       # @example Configured using external config object (~/.fog ignored completely)
       #   @config = Fog::Example::Config.new(...)
       #   @service = Fog::Compute::Example.new(@config)
@@ -100,20 +90,12 @@ module Fog
         setup_requirements
 
         svc = service
-        if Fog.mocking?
-          while svc != Fog::Service
-            service::Mock.send(:include, svc::Collections)
-            svc = svc.superclass
-          end
-          service::Mock.new(cleaned_settings)
-        else
-          while svc != Fog::Service
-            service::Real.send(:include, svc::Collections)
-            svc = svc.superclass
-          end
-          service::Real.send(:include, service::NoLeakInspector)
-          service::Real.new(cleaned_settings)
+        while svc != Fog::Service
+          service::Real.send(:include, svc::Collections)
+          svc = svc.superclass
         end
+        service::Real.send(:include, service::NoLeakInspector)
+        service::Real.new(cleaned_settings)
       end
 
       # @deprecated
@@ -136,7 +118,6 @@ module Fog
 
         require_models
         require_collections_and_define
-        require_requests_and_mock
         @required = true
       end
 
@@ -176,10 +157,6 @@ module Fog
                            end
           end
         end
-      end
-
-      def mocked_requests
-        @mocked_requests ||= []
       end
 
       def model(new_model, path = nil)
@@ -293,22 +270,6 @@ module Fog
       # @return [String] in camel case
       def camel_case_collection_name(collection)
         collection.to_s.split("_").map(&:capitalize).join
-      end
-
-      # This will attempt to require all request files declared in the service using fog"s DSL
-      def require_requests_and_mock
-        requests.each do |request|
-          require_item(request, @request_path)
-          if service::Mock.method_defined?(request.last)
-            mocked_requests << request.last
-          else
-            service::Mock.module_eval <<-EOS, __FILE__, __LINE__
-              def #{request.last}(*args)
-                Fog::Mock.not_implemented
-              end
-            EOS
-          end
-        end
       end
 
       # Requires the correct file for an item (collection, model, or request).
